Fixed warnings from VC++ 6.0.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@69476 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/run-time/rout_obj.c
@@ -325,8 +325,8 @@ rt_public void rout_obj_call_procedure_dynamic (
 	EIF_REFERENCE open_map)
 {
 	EIF_GET_CONTEXT
-	size_t i = 2;
-	size_t args_count = open_count + closed_count;
+	int i = 2;
+	int args_count = open_count + closed_count;
 	int next_open = 0xFFFF;
 	int open_idx = 1;
 	int closed_idx = 1;
